Extract `{Crate,}Bootstrap` bootstrap self-test steps into `bootstrap_self_tests` module
